User sees an error message when there's a networking error.
